Allow injection of optional fields for Message
Testing out an approach for allowing optional fields for a message.
Right now, we're patching the Message class to allow for our own
custom logic for setting the destination property, as well as for
some additional custom headers that we want to treat as properties.

Ideally, we'd be able to inject this behavior without overwriting
additional properties. This is mostly a test to see if this approach
works for our use case and can get something up and running.